Changed type of training point to double. Wrote PshStatistic with several useful statistics. Removed SemanticIndividual - it is not needed here at the moment. Changing version to 0.0.3

+++ src/main/java/org/pushpmx/problem/Keijzer4Problem.java
@@ -3,9 +3,9 @@ package org.pushpmx.problem;
 public class Keijzer4Problem extends FloatSymbolicRegression {
 
 	@Override
-	protected float evaluateFunction(float x) {
-		return (float) ((x * x * x) * Math.exp(-x) * Math.cos(x) * Math.sin(x) * (Math
-				.sin(x) * Math.sin(x) * Math.cos(x) - 1.0f));
+	protected double evaluateFunction(double x) {
+		return ((x * x * x) * Math.exp(-x) * Math.cos(x) * Math.sin(x) * (Math
+				.sin(x) * Math.sin(x) * Math.cos(x) - 1.0));
 	}
 
 }
